Expose only injection APIs when ENABLE_JSON_RPC_API = false
